requirements.txt
django~=3.2.13
djangorestframework~=3.13.1
django-filter~=21.1
celery~=5.2.6